+++ src/components/Preview/components/Settings.vue
@@ -32,6 +32,33 @@
 					/>
 				</v-col>
 			</v-row>
+			<v-row
+				elevation="0"
+				class="d-flex white rounded justify-center align-center pa-4 mt-4"
+			>
+				<v-col
+					cols="12"
+					md="6"
+					class="pa-0 ma-0"
+				>
+					Proxy Images
+				</v-col>
+				<v-col
+					cols="12"
+					md="6"
+					class="pa-0 ma-0"
+				>
+					<v-select
+						v-model="proxyImagesValue"
+						class="right-input"
+						dense
+						outlined
+						label="Should proxy?"
+						:items="[true, false]"
+						:hide-details="true"
+					/>
+				</v-col>
+			</v-row>
 		</OptionWrapper>
 	</div>
 </template>
@@ -47,6 +74,7 @@ export default {
 	computed: {
 		...mapState({
 			defaultView: state => state.previewConfig.settings.defaultView,
+			proxyImages: state => state.previewConfig.settings.proxyImages,
 		}),
 		defaultViewValue: {
 			get() {
@@ -57,10 +85,20 @@ export default {
 				this.updatePreviewSettings();
 			},
 		},
+		proxyImagesValue: {
+			get() {
+				return this.proxyImages;
+			},
+			set(value) {
+				this.setProxyImagesOnPreview(value);
+				this.updatePreviewSettings();
+			},
+		},
 	},
 	methods: {
 		...mapMutations({
 			setDefaultView: "setDefaultView",
+			setProxyImagesOnPreview: "setProxyImagesOnPreview",
 		}),
 		...mapActions({
 			updatePreviewSettings: "updatePreviewSettings",